+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use NamedTuple to represent result of key derivation with Argon2
+- Use Pydantic's Secret types when dealing with passwords. This prevents the data from being printed in the logs and tracebacks.
 
 ### Fixed